requirements.txt
pytest
pytest-mock
python-dateutil
boto3
rope